cmd/test: added --keep_runtime to keep the resulting k8s namespace.

+++ internal/testing/run.go
@@ -27,7 +27,8 @@ import (
 )
 
 type testRun struct {
-	Env ops.WorkspaceEnvironment // Doesn't affect the output.
+	Env            ops.WorkspaceEnvironment // Doesn't affect the output.
+	CleanupRuntime bool                     // Doesn't affect the output.
 
 	TestName       string
 	TestBinPkg     schema.PackageName
@@ -63,13 +64,15 @@ func (rt *testRun) Inputs() *compute.In {
 func (rt *testRun) Compute(ctx context.Context, r compute.Resolved) (fs.FS, error) {
 	p := compute.GetDepValue(r, rt.Plan, "plan")
 
-	compute.On(ctx).Cleanup(tasks.Action("test.cleanup"), func(ctx context.Context) error {
-		if err := runtime.For(rt.Env).DeleteRecursively(ctx); err != nil {
-			return err
-		}
+	if rt.CleanupRuntime {
+		compute.On(ctx).Cleanup(tasks.Action("test.cleanup"), func(ctx context.Context) error {
+			if err := runtime.For(rt.Env).DeleteRecursively(ctx); err != nil {
+				return err
+			}
 
-		return nil
-	})
+			return nil
+		})
+	}
 
 	waiters, err := p.Deployer.Apply(ctx, runtime.TaskServerDeploy, rt.Env)
 	if err != nil {
